rename BacklogItemPlanningPanel etc.
